Reorganised morris sample folders to contain similar code as modules
